requirements.txt
amqp==5.1.0
asgiref==3.4.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
bs4==0.0.1
cached-property==1.5.2
celery==5.1.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.12
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.8.3
kombu==5.1.0
lxml==4.8.0
prompt-toolkit==3.0.28
pytz==2022.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
typing_extensions==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
zipp==3.6.0
